+++ emwiki/article/static/article/js/app.js
@@ -1,7 +1,7 @@
 import {router} from './router/index.js';
 import {ArticleDrawer} from './components/ArticleDrawer.js';
 import {TheoremDrawer} from './components/TheoremDrawer.js';
-import {context} from './context.js';
+import {context} from '../../js/context.js';
 
 new Vue({
   el: '#app',
