+++ emwiki/article/static/article/js/models/parser.js
@@ -1,4 +1,5 @@
 import {Comment} from './comment.js';
+
 /**
  * Parser
  */
