+++ emwiki/graph/static/graph/js/app.js
@@ -1,6 +1,6 @@
 import {GraphDrawer} from './components/GraphDrawer.js';
 import {GraphService} from './services/graph-service.js';
-import {context} from './context.js';
+import {context} from '../../js/context.js';
 
 new Vue({
   el: '#app',
