--- emwiki/symbol/static/symbol/js/context.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line max-len
-export const context = JSON.parse(document.getElementById('context').textContent);
